pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rag-news"
version = "0.1.0"
description = "Self-correcting news RAG with LangGraph, Tavily, ChromaDB, and Telegram"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "apscheduler>=3.10,<4.0",
  "chromadb>=0.5,<1.0",
  "fastapi>=0.115,<1.0",
  "google-genai>=1.0,<2.0",
  "httpx>=0.27,<1.0",
  "langgraph>=0.2,<0.4",
  "openai>=1.40,<2.0",
  "posthog>=2.4,<6.0",
  "pydantic>=2.7,<3.0",
  "python-telegram-bot>=21.0,<22.0",
  "python-dotenv>=1.0,<2.0",
  "sentence-transformers>=3.0,<4.0",
  "tavily-python>=0.5,<1.0",
  "uvicorn[standard]>=0.30,<1.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0",
  "pytest-asyncio>=0.23",
  "ruff>=0.6",
]

[project.scripts]
rag-news = "rag_news.interfaces.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-q"
asyncio_mode = "auto"
